refactor: split the planner's nodes, rules, and lowering by owner
nodes.rs, rules.rs, and planner.rs each held every index type's contribution
at once, so adding an index type meant editing three shared files and reading
past two other index types to do it.

Reorganize so the framework is split by stage and each index type keeps its
own node, rules, and lowering together: vector/ and fts/ now reach the
framework through the same five entry points, and take/ holds late
materialization. Cross-index concerns that no single index can decide stay
central: rule ordering in mod.rs, and the coverage split — which serves every
search node type through the SplittableSearch trait — in coverage.rs.

Largest module drops from 1790 to 501 lines. No behavior change.
