+++ rust/lance/src/dataset/scanner/logical/take/mod.rs
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright The Lance Authors
+
+//! Late materialization: fetching the columns a search did not produce.
+
+mod node;
+mod planner;
+
+pub use node::*;
+pub use planner::*;
